project_final/requirements.txt
dagster==1.5.12
dagster-webserver==1.5.12
faker==20.1.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.4
numpy>=1.25.0
python-dotenv==1.0.0
schedule==1.2.0
